refactor: base complete for unit tests
