+++ tests/mod.ts
@@ -1,0 +1,14 @@
+// THE ORDER OF THE IMPORTS IN THIS FILE MATTER!
+// DO NOT MOVE THEM UNLESS YOU KNOW WHAT YOUR DOING!
+
+// First complete non-api reliant testing. 
+// Don't waste api rate limits if a early test fails.
+import "./util/utils.ts";
+
+// API TESTING BELOW
+
+// First initate the connection
+import "./ws/start_bot.ts";
+import "./guilds/create_guild.ts";
+import "./guilds/delete_server.ts";
+import "./ws/ws_close.ts";
